move custom dp class under sd modeling
