+++ optimum/neuron/utils/misc.py
@@ -50,8 +50,7 @@ from .require_utils import (
 
 
 if is_torch_neuronx_available():
-    from torch_neuronx import DataParallel, move_trace_to_device
-    from torch_neuronx.experimental.placement import set_neuron_cores
+    from torch_neuronx import DataParallel
 
 if TYPE_CHECKING:
     from transformers.modeling_utils import PreTrainedModel
@@ -705,29 +704,3 @@ def get_stable_diffusion_configs(
             configs[name] = models_for_export[name].config
 
     return configs
-
-
-# TO REMOVE: This class will be included directly in the DDP API of Neuron SDK 2.20
-class WeightSeparatedDataParallel(DataParallel):
-
-    def _load_modules(self, module):
-        try:
-            self.device_ids.sort()
-
-            loaded_modules = [module]
-            # If device_ids is non-consecutive, perform deepcopy's and load onto each core independently.
-            for i in range(len(self.device_ids) - 1):
-                loaded_modules.append(copy.deepcopy(module))
-            for i, nc_index in enumerate(self.device_ids):
-                set_neuron_cores(loaded_modules[i], nc_index, 1)
-                move_trace_to_device(loaded_modules[i], nc_index)
-
-        except ValueError as err:
-            self.dynamic_batching_failed = True
-            logger.warning(f"Automatic dynamic batching failed due to {err}.")
-            logger.warning(
-                "Please disable dynamic batching by calling `disable_dynamic_batching()` "
-                "on your DataParallel module."
-            )
-        self.num_workers = 2 * len(loaded_modules)
-        return loaded_modules
